Tests/LibWeb/Text/expected/css/FontFace-arraybuffer-matching.txt

face.status: loaded
font matched: true
font unmatched after delete: true
